+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add planting and harvest year to proposals. [#481](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/481)
+
 ### Fixed
 
 - Correct data steward field reference to data curator researcher. [#445](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/445)